-> added model classes -> added User, Role, Permission, ContactInfos, AddressInfos

+++ users-microservice/src/main/java/com/bt/dev/kodemy/users/model/Permission.java
@@ -16,7 +16,7 @@ public class Permission {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "permission_id")
-    private Integer permissionId;
+    private Long permissionId;
 
     @Column(name = "permission", nullable = false)
     private String permission;
